+++ CHANGELOG.md
@@ -13,6 +13,9 @@
     Click to see more.
   </summary>
   
+  ### :bug: Bug Fix
+  - `divider`
+    - [#511](https://github.com/wix-incubator/rich-content/pull/511) divider's width set to 100% in editor&viewer when no initial state of width is given
   ### :house: Internal
   - `editor`
     - [#510](https://github.com/wix-incubator/rich-content/pull/510) custom icons code refactoring